IB/hfi1: Fix static checker issues in affinity.c
This change fixes 3 klocwork issues in affinity.c.

Two are alloc/free ordering issues with cpumask_var_t.

The other is a false positive.  It is two variable dependency
that klockwork does not recognize.  In this case: cpu is -1,
therefore set is assigned a valid value.

INTERNAL: PR 130673 - Driver crash running MPI-Stress-Psm on fpga-1,4

Reviewed-by: Dennis Dalessandro <dennis.dalessandro@intel.com>
Signed-off-by: Dean Luick <dean.luick@intel.com>
Change-Id: I22a21c5f4201895ae3526110bc600b5315ca96a4