IB/hfi1: Tune for unknown channel if configuration file is absent
Currently, the driver fails to tune the SerDes and therefore prevents
link up if the configuration file is missing or fails parsing or
validation. This patch adds a fallback option so that the 8051 is asked
to tune for an unknown channel and possibly get the link up if tuning
succeeds. It also adds a user-friendly message to update the
configuration file if it is out-of-date.

INTERNAL: PR 130956 - WFR driver should fallback to unknown channel if
INTERNAL: config file is unavailable
INTERNAL: SL-786 05: 130956: WFR driver should fallback to unknown
INTERNAL: channel if config file unavailable

Reviewed-by: Mike Marciniszyn <mike.marciniszyn@intel.com>
Reviewed-by: Dean Luick <dean.luick@intel.com>
Signed-off-by: Easwar Hariharan <easwar.hariharan@intel.com>
Change-Id: I996589f2be58c6a1bd63ce274fe82a85e4039383