STL-3298 correctly set ibmtu and QP's path mtu
The path_mtu in the QP structure was incorrect. This is needed for UC queue
pairs and is caused by the port structure's ibmtu not being set.  This should
be set to the maximum of the per-VL MTUs (based on usage in modify_qp).
Some other various MTU cleanup in this commit, mainly PCI driver must maintain
per-VL MTUs separately for each port.

Reviewed-by: Sudeep Dutt <sudeep.dutt@intel.com>
Reviewed-by: Sanath Kumar <sanath.s.kumar@intel.com>
Signed-off-by: Brian Welty <brian.welty@intel.com>