IB/hfi1: Insure last cursor is updated prior to complete
This closes a race between a poll cq seeing a complete and
a post send checking for a full send queue introduced by
Commit bfc5da43b771 ("IB/hfi1: add s_hlock for use in post send").

All completion paths insure the s_last is visible prior to storing the
completion.  The smp memory barrier is not needed since storing a
completion has one, but a compiler barrier is used to insure the
s_last update isn't moved. rc completion paths insure the over
increment in the wrap case is never visible to the post send by
doing the wrap in an automatic.

INTERNAL: PR 131281 - ibv_post_send occasional returns ENOMEM

Change-Id: I0e227475c89fd20eff09323596f9391101893e50
Reviewed-by: Dennis Dalessandro <dennis.dalessandro@intel.com>
Signed-off-by: Mike Marciniszyn <mike.marciniszyn@intel.com>