STL-189 Use GENERAL_DMA for STLEEP-EP as well
EEPH expects the hardware to insert the CRC but reserves
space for the ICRC and tail bytes. Use GENERAL_DMA for EEPH
as well instead of GENERAL_DMA_NOCRC but subtract the ICRC,
padding and tail bytes from the length so that the Simics
model adds that.

There is still an arch open about FXR GENERAL_DMA handling so this
implementation might need some tweaking once a decision is made there.

Change-Id: Ic92325c1766df070fe1719218cc3fb85a2776863
Reviewed-by: Niranjana Vishwanathapura <niranjana.vishwanathapura@intel.com>
Signed-off-by: Sudeep Dutt <sudeep.dutt@intel.com>