STL-3700 Correct the receive buffer size to 128KB
The constant used earlier (BUFF_CREDIT_LIMIT ~= 1MB) was incorrect.
This resulted in FM programming the credit managers with
credits greater than 128KB (= 2048 credits for vAU of 3).
This patch hardcodes the receive size and calculates
the credits based on the vAU constant.
Thanks to Miao and Matt for reporting this problem

Change-Id: Icb39dfcbf61e23bf2a6423be4458fb5d15ca2811
Reviewed-by: Brian Welty <brian.welty@intel.com>
Reviewed-by: Sudeep Dutt <sudeep.dutt@intel.com>
Signed-off-by: Sanath Kumar <sanath.s.kumar@intel.com>