IB/hfi1: Set SDMA slot status on error
The SDMA slot status was not being set if the packet
generation fuction returned an error. That left the
slot in the "used" state, effectively leaking it.

This slot leak would eventually result in an error
being returned to PSM due to the slot being in the
"used" state when it should not have been.

Fix this leak by properly setting the slot to the
ERROR state.

INTERNAL: PR 131550 - mpistress gives 'writev(): Invalid
ITNERNAL: slot' during low ulimit settings

Reviewed-by: Dennis Dalessandro <dennis.dalessandro@intel.com>
Signed-off-by: Mitko Haralanov <mitko.haralanov@intel.com>
Change-Id: Ic27f51ce6937b153e9da1cf399387acc16b63a2b