STL-3334 complete user QP support (2 of 2)
Fix how IOVEC array allocation was being associated with active WQE.
For FIRST+MIDDLE(s)+LAST, we need multiple IOVEC arrays.  Change allocation
to use a new structure that associates the temporary IOVEC array with WQE.

Unfortunately we now have to make a copy of the IB header used for each IOVEC
array as the MIDDLE and LAST packets will overwrite the IB header in the QP.
This could be optimized to build the header in new memory location but would
require refactoring upper layer code, so should wait for RDMA-VT integration.

Reviewed-by: Sudeep Dutt <sudeep.dutt@intel.com>
Signed-off-by: Brian Welty <brian.welty@intel.com>