IB/core: correct issue with sge copy in corrupting wr
Commit e622f2f4ad21 ("IB: split struct ib_send_wr")
introduced a regression for HCAs whose user mode post
sends go through ib_uverbs_post_send().

The code didn't account for the fact that the first sge is
offset by an operation dependent length.  The allocation did,
but the pointer to the sge list is computed without that knowledge.

Store the operation dependent length in an automatic and
compute the sge list copy in destination using that length.

INTERNAL: PR 131767 - 4.4.0-rc2 qib is broken

Change-Id: Ida924231563a5784781d42153d69a14979bb9d2f
Reviewed-by: Ira Weiny <ira.weiny@intel.com>
Signed-off-by: Mike Marciniszyn <mike.marciniszyn@intel.com>