IB/hfi1: Fix for module parameter rcvhdrcnt when it's 2097152
The driver crashes when loaded with parameter rcvhdrcnt=2097152.
The root cause was that rcvhdrcnt was initially a 32 bit variable
and its value was assign to a 16 bit variable, truncating the
upper 16 bits. This patch prevents the user from passing a value
for rcvhdrcnt greater than 16352 (Maximum number for rcvhdrcnt).

INTERNAL: PR 131689 - Crash and reboot when loading hfi1
INTERNAL: with rcvhdrcnt=2097152
INTERNAL: SL-991 - 011:131689:Crash and reboot when
INTERNAL: loading hfi1 with rcvhdrcnt=2097152

Change-Id: Ifcb6adf851d5500c54d0a9aa635fb49210fcd2e7
Reviewed-by: Dean Luick <dean.luick@intel.com>
Reviewed-by: Haralanov, Mitko <mitko.haralanov@intel.com>
Signed-off-by: Sebastian Sanchez <sebastian.sanchez@intel.com>