IB/hfi1: Fix bug with RcvHrdQ Tail update setting
Commit 67639a40231ab132fa24b0b51c502fa555db1b3b (IB/hfi1:
Workaround to prevent corruption during packet delivery)
introduced a bug with the value of the TailUpd bit of the
RcvCtxtCtrl register.

When the context was disabled for the first time and the
RcvHdrTailAddr value was set to the dummy address, the value
of the TailUpd bit was set to 1 instead of cleared. This
value propagated to any subsequent uses of that context,
causing degraded performance.

This was made even worse by the fact that the user context
were not properly reset when being closed.

This commit addresses both of the above issues.

INTERNAL: PR 131999 - LLNL: MPI/PSM message rate drop-off for
INTERNAL: 16B messages seen for CTS-1
INTERNAL: SL-1002 000: 131999: LLNL: MPI/PSM message rate
INTERNAL: drop-off for 16B messages seen for CTS-1

Reviewed-by: Dennis Dalessandro <dennis.dalessandro@intel.com>
Signed-off-by: Mitko Haralanov <mitko.haralanov@intel.com>
Change-Id: I199acc52172bcefffbb584f25afd3ae75c5c332c