STL-1901 Apply SMA changes from wfr-driver
Since SMA is now part of the PCIe driver, the code flow now is similar
to that of WFR. So this patch makes SMA code common between
WFR and FXR as similar as possible. Hopefully it will be
easier to diff in the future if we decide to have common SMA
layer in RDMA-VT (for ex). Also part of this patch brings in
any bug fixes and missed out code to the FXR SMA.

WFR driver code referred:
wfr-driver repo, for-upstream branch and commit
c2ef7b86a3b08851e773306871c8e0d6807b9cea

Change-Id: I6ecd6d888feaebfa9652b045926040263880e615
Reviewed-by: Sudeep Dutt <sudeep.dutt@intel.com>
Reviewed-by: Brian Welty <brian.welty@intel.com>
Signed-off-by: Sanath Kumar <sanath.s.kumar@intel.com>