IB/hfi1: Use device file minor to identify EPROM
When writing to the EPROM, the driver will always use the
"first" device.  This is incorrect for multiple cards.

Use the device file minor to determine the device to use.
Reject the generic device file.

INTERNAL: PR 132104 - hfi1_eprom utility only supports writing to a
INTERNAL: second CHF through the direct CSR interface

Reviewed-by: Mitko Haralanov <mitko.haralanov@intel.com>
Signed-off-by: Dean Luick <dean.luick@intel.com>
Change-Id: I0098c26a90ef402fd7374452555091005e34993e