IB/hfi1: fix for crash due to null pointer dereference
The test to ensure that a packet is built is to check if
s_hdrwords is not zero. But there are some cases of txreq allocation
failures where s_hdrwords is not set to zero. This was the cause of the
null pointer dereference where txreq was null and yet being derefenced.

The fix was to set s_hdrwords to zero in cases where txreq was NULL.
This avoids the null pointer dereference.

INTERNAL: PR 132126 - kernel crash - BUG: unable to handle kernel NULL
INTERNAL: pointer dereference
SL-808 - Ensure header words is 0 when txreq is NULL

Change-Id: I5822b61a7ecd76cdae69df1856a51ea3533e1a6a
Reviewed-by: Mike Marciniszyn <mike.marciniszyn@intel.com>
Reviewed-by: Mitko Haralanov <mitko.haralanov@intel.com>
Signed-off-by: Vennila Megavannan <vennila.megavannan@intel.com>