staging/rdma/hfi1: Remove driver specific members from hfi1 qp type
In preparation for moving the queue pair data structure to rdmavt the
members of the driver specific queue pairs which are not common need to be
pushed off to a private driver structure. This structure will be available
in the queue pair once moved to rdmavt as a void pointer. This patch while
not adding a lot of value in and of itself is a prerequisite to move the
queue pair out of the drivers and into rdmavt.

The driver specific, private queue pair data structure should condense as
more of the send side code moves to rdmavt.

Reviewed-by: Mike Marciniszyn <mike.marciniszyn@intel.com>
Signed-off-by: Dennis Dalessandro <dennis.dalessandro@intel.com>

---

Conflicts:
	drivers/staging/rdma/hfi1/ruc.c
	drivers/staging/rdma/hfi1/ud.c
	drivers/staging/rdma/hfi1/verbs.c
	drivers/staging/rdma/hfi1/verbs.h

		Major conflicts here...  This patch leaves in the conflicts
		commented out but made a lot of other changes...

		Update: I believe all the code from upstream is properly merged
		at this time.

TO BE SQUASHED: ***WIP*** Remove driver specific members from qp
	place r_adefered into priv

TO BE SQUASHED: Add back in the optimization from

commit 13a69f5299688074666e1f74b90ddc6beae43e6d
Author: Mike Marciniszyn <mike.marciniszyn@intel.com>
Date:   Mon Oct 26 10:28:27 2015 -0400

    staging/rdma/hfi1: Fix regression in send performance

    The clear_ahg call is new in hfi1 vs qib.  For small messages the progress
    routine always builds one and clears out the ahg state when the queue has gone
    to empty which is the predominant case for small messages.

    Inline the routine and avoid the call to sdma_ahg_free to mitigate the
    performance impact.  Finally, move the routine to qp.h for scope reasons.

    Reviewed-by: Dennis Dalessandro <dennis.dalessandro@intel.com>
    Signed-off-by: Mike Marciniszyn <mike.marciniszyn@intel.com>
    Signed-off-by: Ira Weiny <ira.weiny@intel.com>
    Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>

	This was
	+       if (qp->s_sde && qp->s_ahgidx >= 0)

TO BE SQUASHED: sde should be a qp_priv member
	Make s_sde private to hfi1