staging/rdma/hfi1: Use rdmavt send flags and recv flags
Use the definitions of the s_flags and r_flags which are now in rdmavt.

Reviewed-by: Ira Weiny <ira.weiny@intel.com>
Reviewed-by: Harish Chegondi <harish.chegondi@intel.com>
Signed-off-by: Dennis Dalessandro <dennis.dalessandro@intel.com>

Conflicts:
	drivers/staging/rdma/hfi1/driver.c
	drivers/staging/rdma/hfi1/qp.c
	drivers/staging/rdma/hfi1/rc.c
	drivers/staging/rdma/hfi1/ruc.c
	drivers/staging/rdma/hfi1/verbs.h
		Major conflicts here as well.

		I particularly worry about the delayed Ack code which was
		upstream which caused conflicts.

		Added back in qp->s_ahgidx to clear_ahg

TO BE SQUASHED: remove conflict comments

We are going to use the RVT_R_RSP_NAK bit instead of the DEFERED_ACK bit...

	Even though it is a bad name we are just going to use the "standard"
	bit of RVT_R_RSP_NAK.

TO BE SQUASHED: move migrate_qp back where it belongs.

The original patch did not have this move of migrate qp so don't move it.

Also put these lines back...

+       priv->s_sc = ah_to_sc(qp->ibqp.device, &qp->remote_ah_attr);
+       priv->s_sde = qp_to_sdma_engine(qp, priv->s_sc);