STL-4368 Fix EEPH hang when packet with invalid MTU is transmitted
Added checks to:
a) ensure packets larger than the max MTU are not TX'd
b) ensure eager buffer length received in the EQ is not
   larger than the max MTU

Simics does not support transmitting packets greater than 8192 bytes
but this patch also increases the eager buffer length to 16K so that
MAX MTU of 10240 can be supported once Simics support is enabled.
The eager buffer size has to be a power of two and hence the 16K.

Tested that sending packets greater than the MTU fails gracefully
with an error instead of crashing the driver.

Change-Id: I905f07cf8ab6db6d95de1d3ab46bc0e452ec6a00
Reviewed-by: Brian Welty <brian.welty@intel.com>
Reviewed-by: Niranjana Vishwanathapura <niranjana.vishwanathapura@intel.com>
Signed-off-by: Sudeep Dutt <sudeep.dutt@intel.com>