STL-4089 implement API for setting RSM rules
Approach here is for driver to manage assignment of the RSM_MAP enties
via a bitmap.  This makes usage dynamic and modifying a single RSM rule
later should potentially not affect other RSM users.
Receive contexts must be setup first, and then RSM rule is configured
with a supplied array of hfi_ctxs to use for the RSM_MAP entries.
Tested with 16B verbs rule which will posted as separate change.

Change-Id: I68f7e28ec42ef4d046ac2b446dfb7480fd60e156
Reviewed-by: Niranjana Vishwanathapura <niranjana.vishwanathapura@intel.com>
Reviewed-by: Sudeep Dutt <sudeep.dutt@intel.com>
Signed-off-by: Brian Welty <brian.welty@intel.com>