IB/hfi1: Fix header size calculation for RC/UC QPs with GRH enabled
There is a header size counter in both the QP struture and the txreq
structure. The counter in the txreq structure is not updated properly
for RC and UC queue pairs with GRH enabled, and thus causing SDMA
send to fail.

The UD path has the counter updated properly, but leaves some
redundant assignments.

This patch fixes the RC and UC path and removes the redundant
assignments from the UD path.

INTERNAL: PR 131613 - RC QP verbs with global headers does not appear
INTERNAL: to work
INTERNAL: SL-1197

Reviewed-by: Dean Luick <dean.luick@intel.com>
Reviewed-by: Mike Marciniszyn <mike.marciniszyn@intel.com>
Signed-off-by: Jianxin Xiong <jianxin.xiong@intel.com>
Change-Id: Icae0c1ff9863536eb5b4d8ff6512526ff029e5df